rename numpy connector
